# Download Hugging Face models into Hugging Face cache