--- a/torchchat/cli/download.py
+++ b/torchchat/cli/download.py
@@ -18,15 +18,19 @@ from torchchat.model_config.model_config import (
     resolve_model_config,
 )
 
+# By default, download models from HuggingFace to the Hugginface hub directory.
+# Both $HF_HOME and $HUGGINGFACE_HUB_CACHE are valid environment variables for the same directory.
+HUGGINGFACE_HOME_PATH = Path(os.environ.get("HF_HOME", os.environ.get("HUGGINGFACE_HUB_CACHE", os.path.expanduser("~/.cache/huggingface/hub"))))
 
 def _download_hf_snapshot(
-    model_config: ModelConfig, artifact_dir: Path, hf_token: Optional[str]
+    model_config: ModelConfig, hf_token: Optional[str]
 ):
     from huggingface_hub import model_info, snapshot_download
     from requests.exceptions import HTTPError
 
     # Download and store the HF model artifacts.
-    print(f"Downloading {model_config.name} from HuggingFace...", file=sys.stderr)
+    model_dir = get_model_dir(model_config, None)
+    print(f"Downloading {model_config.name} from Hugging Face to {model_dir}", file=sys.stderr, flush=True)
     try:
         # Fetch the info about the model's repo
         model_info = model_info(model_config.distribution_path, token=hf_token)
@@ -56,8 +60,6 @@ def _download_hf_snapshot(
 
         snapshot_download(
             model_config.distribution_path,
-            local_dir=artifact_dir,
-            local_dir_use_symlinks=False,
             token=hf_token,
             ignore_patterns=ignore_patterns,
         )
@@ -76,16 +78,20 @@ def _download_hf_snapshot(
         else:
             raise e
 
+    # Update the model dir to include the snapshot we just downloaded. 
+    model_dir = get_model_dir(model_config, None)
+    print("Model downloaded to", model_dir)
+
     # Convert the Multimodal Llama model to the torchtune format.
     if model_config.name in {"meta-llama/Llama-3.2-11B-Vision-Instruct", "meta-llama/Llama-3.2-11B-Vision"}:
         print(f"Converting {model_config.name} to torchtune format...", file=sys.stderr)
-        convert_hf_checkpoint_to_tune( model_dir=artifact_dir, model_name=model_config.name)
+        convert_hf_checkpoint_to_tune( model_dir=model_dir, model_name=model_config.name)
 
     else:
         # Convert the model to the torchchat format.
         print(f"Converting {model_config.name} to torchchat format...", file=sys.stderr)
         convert_hf_checkpoint(
-            model_dir=artifact_dir, model_name=model_config.name, remove_bin_files=True
+            model_dir=model_dir, model_name=model_config.name, remove_bin_files=True
         )
 
 
@@ -99,12 +105,51 @@ def _download_direct(
         print(f"Downloading {url}...", file=sys.stderr)
         urllib.request.urlretrieve(url, str(local_path.absolute()))
 
+def _get_hf_artifact_dir(model_config: ModelConfig) -> Path:
+    """
+    Returns the directory where the model artifacts are stored.
+    
+    This is the root folder with blobs, refs and snapshots
+    """
+    assert(model_config.distribution_channel == ModelDistributionChannel.HuggingFaceSnapshot)
+    return HUGGINGFACE_HOME_PATH / f"models--{model_config.distribution_path.replace('/', '--')}"
+
+
+def get_model_dir(model_config: ModelConfig, models_dir: Optional[Path]) -> Path:
+    """
+    Returns the directory where the model artifacts are stored.
+    For HuggingFace snapshots, this is the HuggingFace cache directory.
+    For all other distribution channels, we use the models_dir. 
+    
+    For CLI usage, pass in args.model_directory. 
+    """
+    if model_config.distribution_channel == ModelDistributionChannel.HuggingFaceSnapshot:
+        artifact_dir = _get_hf_artifact_dir(model_config) 
+        
+        # If these paths doesn't exist, it means the model hasn't been downloaded yet.
+        if not os.path.isdir(artifact_dir) and not os.path.isdir(artifact_dir / "snapshots"):
+            return artifact_dir
+        snapshot = open(artifact_dir / "refs" / "main", "r").read().strip()
+        return artifact_dir / "snapshots" / snapshot 
+    else:
+        return models_dir / model_config.name
+
 
 def download_and_convert(
     model: str, models_dir: Path, hf_token: Optional[str] = None
 ) -> None:
     model_config = resolve_model_config(model)
-    model_dir = models_dir / model_config.name
+    model_dir = get_model_dir(model_config, models_dir)
+
+    # HuggingFace download
+    if (
+        model_config.distribution_channel
+        == ModelDistributionChannel.HuggingFaceSnapshot
+    ):
+        _download_hf_snapshot(model_config, hf_token)
+        return
+
+    # Direct download
 
     # Download into a temporary directory. We'll move to the final
     # location once the download and conversion is complete. This
@@ -117,11 +162,6 @@ def download_and_convert(
 
     try:
         if (
-            model_config.distribution_channel
-            == ModelDistributionChannel.HuggingFaceSnapshot
-        ):
-            _download_hf_snapshot(model_config, temp_dir, hf_token)
-        elif (
             model_config.distribution_channel == ModelDistributionChannel.DirectDownload
         ):
             _download_direct(model_config, temp_dir)
@@ -144,9 +184,9 @@ def download_and_convert(
 
 def is_model_downloaded(model: str, models_dir: Path) -> bool:
     model_config = resolve_model_config(model)
-
+    
     # Check if the model directory exists and is not empty.
-    model_dir = models_dir / model_config.name
+    model_dir = get_model_dir(model_config, models_dir)
     return os.path.isdir(model_dir) and os.listdir(model_dir)
 
 
@@ -194,13 +234,16 @@ def remove_main(args) -> None:
         return
 
     model_config = resolve_model_config(args.model)
-    model_dir = args.model_directory / model_config.name
+    model_dir = get_model_dir(model_config, args.model_directory)
 
     if not os.path.isdir(model_dir):
-        print(f"Model {args.model} has no downloaded artifacts.")
+        print(f"Model {args.model} has no downloaded artifacts in {model_dir}.")
         return
+    if model_config.distribution_channel == ModelDistributionChannel.HuggingFaceSnapshot:
+        # For HuggingFace models, we need to remove the entire root directory.
+        model_dir = _get_hf_artifact_dir(model_config)
 
-    print(f"Removing downloaded model artifacts for {args.model}...")
+    print(f"Removing downloaded model artifacts for {args.model} at {model_dir}...")
     shutil.rmtree(model_dir)
     print("Done.")
 
@@ -216,10 +259,10 @@ def where_main(args) -> None:
         return
 
     model_config = resolve_model_config(args.model)
-    model_dir = args.model_directory / model_config.name
+    model_dir = get_model_dir(model_config, args.model_directory)
 
     if not os.path.isdir(model_dir):
-        raise RuntimeError(f"Model {args.model} has no downloaded artifacts.")
+        raise RuntimeError(f"Model {args.model} has no downloaded artifacts in {model_dir}.")
 
     print(str(os.path.abspath(model_dir)))
     exit(0)