@byjlw 2024-10-14
> Fixing @swolchok 's suggestions in follow-up commits.
> 
> Regarding this:
> 
> > I don't see any kind of migration from the old path to the new one. How are people going to know we've abandoned gigabytes of downloaded models? Can we check the old path and move models over?
> 
> What's the best way to do this? I've currently implemented this as:
> 
> * User calls download
> * We check _all model configs_ where the distribution is through HF.
> * If that model exists in the old location, delete it
> 
> Testing:
> 
> ```
> python3 torchchat.py download llama3.2-3b
> Cleaning up old model artifacts in /Users/puri/.torchchat/model-cache/meta-llama/Llama-2-7b-chat-hf. New artifacts will be downloaded to /Users/puri/.cache/huggingface/hub
> Cleaning up old model artifacts in /Users/puri/.torchchat/model-cache/meta-llama/Meta-Llama-3.1-8B-Instruct. New artifacts will be downloaded to /Users/puri/.cache/huggingface/hub
> Cleaning up old model artifacts in /Users/puri/.torchchat/model-cache/meta-llama/Llama-3.2-11B-Vision-Instruct. New artifacts will be downloaded to /Users/puri/.cache/huggingface/hub
> Downloading meta-llama/Meta-Llama-3.2-3B-Instruct from Hugging Face to /Users/puri/.cache/huggingface/hub/models--meta-llama--Llama-3.2-3B-Instruct/snapshots/392a143b624368100f77a3eafaa4a2468ba50a72
> ...
> ```
> 
> This should clean up everything in one fell-swoop. We could figure out a "lazy" way of doing this, but I think this is best.

Why redownload? 
Copy to new location would be much more efficient. 

Also i recommend updating the description for this PR so it's very clear what we'll do and for how long as well as details on how to change the default the cache location.