@vmpuri 2024-10-16
>Why redownload?
Copy to new location would be much more efficient.

True - this was my first thought, but there were 2 main things that make this more complex (and therefore more error prone). 
1. **Possible Existing Mutations** - It's possible artifacts in the existing ~/.torchchat directory are modified despite having the same name. If we transfer this, then a user attempts to use it from another application, it wouldn't be operating on the original files. We should ensure the Hugging Face cache is read-only. 
2. **Replicating directory structure** I'm not sure how to source all of the hashes and paths for the artifacts or the snapshot ID - I don't believe any of this information is preserved in the current download. Currently, all of the files are exactly as they appear in ~/.torchchat. However, Hugging Face will download the files for a particular snapshot in `<HF_CACHE_DIR>/models--some-model--model-1B/blobs` with some hash as the filename. The proper files symlinked to the blobs, i.e. `<HF_CACHE_DIR>/models--some-model--model-1B/snapshots/<snapshot_id>/original/model.pth`. 


Given that re-downloading the models to the Hugging Face cache is guaranteed to download artifacts to the expected location, I figure the tradeoff between long-term reliability & one-time efficiency makes sense. Re-downloading ~50gb of models would be painful once, but manageable (just ~33 minutes even on a relatively-slow 25 mbps connection).

Let me know your thoughts (or if you know a way to recover the snapshot ID and blob IDs from our current download)